Move model classes to `model` package
