+++ ui-game/src/main/java/com/jayasuryat/uigame/logic/model/GameConfiguration.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jayasuryat.uigame.logic
+package com.jayasuryat.uigame.logic.model
 
 import androidx.compose.runtime.Immutable
